requirements.txt
crewai
crewai-tools
langchain-nvidia-ai-endpoints
langtrace-python-sdk
python-dotenv
pydantic
pyyaml
nest-asyncio